requirements.txt
certifi==2021.10.8
jplephem==2.17
numpy==1.22.3
pyserial==3.5
sgp4==2.20
skyfield==1.42
